Slacktapped v1.0.0 Slacktapped

Functions

handle_checkin(checkin)

Processes a checkin and posts it to Slack, if the checkin is eligible. It adds an “attachments” list to the checkin map, which is used to aggregate individual attachments from the checkin, comments, and badges. The final state of “attachments” is what gets sent directly to the Slack webhook.

Ineligible checkins contain the text “#shh” in the checkin comment. For, you know, day drinking.

Examples

A normal checkin gets through:

iex> Slacktapped.handle_checkin(%{"beer" => %{}})
{:ok,
  %{
    "beer" => %{},
    "reported_as" => "without-image",
    "attachments" => [
      %{
        "author_icon" => nil,
        "author_link" => "https://untappd.com/user/",
        "author_name" => nil,
        "color" => "#FFCF0B",
        "fallback" => " is drinking .",
        "footer" => "<https://untappd.com/brewery/|>",
        "footer_icon" => nil,
        "image_url" => nil,
        "mrkdwn_in" => ["text"],
        "text" => "<https://untappd.com/user/|> is drinking " <>
          "<https://untappd.com/b//|> (, % ABV).\n" <>
          "<https://untappd.com/user//checkin/|Toast »>",
        "title" => nil,
        "title_link" => "https://untappd.com/b//"
      }
    ]
  }
}

A checkin with no beer is ignored:

iex> Slacktapped.handle_checkin(%{})
{:error, %{"attachments" => []}}

Checkins with the text “#shh” in the checkin comment are ignored:

iex> Slacktapped.handle_checkin(%{"checkin_comment" => "#shh"})
{:error, %{"attachments" => [], "checkin_comment" => "#shh"}}

is_eligible_checkin(checkin)

Determines if a checkin is eligible to be processed.

Checkin is ineligible if:

The checkin_comment contains the text “#shh”.

There is no beer (sad).

Examples

iex> Slacktapped.is_eligible_checkin(%{})
{:error, %{}}

iex> Slacktapped.is_eligible_checkin(%{"beer" => %{}})
{:ok, %{"beer" => %{}}}

iex> Slacktapped.is_eligible_checkin(%{"checkin_comment" => "#shh"})
{:error, %{"checkin_comment" => "#shh"}}

main()

Fetches checkins from Untappd and then processes each checkin.

Example

iex> Slacktapped.main :ok

start(type, args)

Callback implementation for Application.start/2.