Slacktapped v1.0.0 Slacktapped.Badges

Functions

is_eligible_badge(badge)

Determines if a badge is eligible to be posted to Slack.

Badge is ineligible if we have already posted it to Slack.

Example

iex> Slacktapped.Badges.is_eligible_badge(%{"user_badge_id" => 5566})
true

iex> Slacktapped.Badges.is_eligible_badge(%{"user_badge_id" => 5533})
false

parse_badge(badge, checkin)

Parses a badge into an attachment for Slack.

Example

iex> Slacktapped.Badges.parse_badge(
...> %{
...>   "badge_name" => "Hopped Up (Level 18)",
...>   "user_badge_id" => 3593,
...>   "badge_image" => %{
...>     "sm" => "http://path/to/badge/image"
...>   }
...> },
...> %{
...>   "attachments" => [],
...>   "user" => %{
...>     "user_name" => "nicksergeant",
...>     "user_avatar" => "http://path/to/user/avatar"
...>   },
...>   "beer" => %{
...>     "bid" => 123,
...>     "beer_name" => "IPA",
...>     "beer_slug" => "two-lake-ipa"
...>   },
...>   "checkin_id" => 567
...> })
%{
  "author_icon" => "http://path/to/user/avatar",
  "author_link" => "https://untappd.com/user/nicksergeant",
  "author_name" => "nicksergeant",
  "color" => "#FFCF0B",
  "fallback" => "nicksergeant earned the badge Hopped Up (Level 18).",
  "image_url" => "http://path/to/badge/image",
  "text" => "<https://untappd.com/user/nicksergeant|nicksergeant> " <>
    "earned the badge " <>
    "<https://untappd.com/user/nicksergeant/badges/3593|" <>
    "Hopped Up (Level 18)> for " <>
    "<https://untappd.com/user/nicksergeant/checkin/567|their checkin> " <>
    "of <https://untappd.com/b/two-lake-ipa/123|IPA>.",
  "title" => "Hopped Up (Level 18)",
  "title_link" => "https://untappd.com/user/nicksergeant/badges/3593"
}

process_badge(badge, checkin)

Reports the badge as posted to Redis, and returns a parsed badge as an attachment for Slack.

process_badges(checkin)

Processes a checkins badges.

Does nothing if the instance is configured to ignore badges, otherwise it loops through all badges and determines if they are eligible to be posted to Slack. If they are, they’re processed and added to the checkins attachments.

Examples

iex> Application.put_env(:slacktapped, :ignore_badges, true)
iex> Slacktapped.Badges.process_badges(%{})
{:error, %{}}

iex> Application.put_env(:slacktapped, :ignore_badges, nil)
iex> Slacktapped.Badges.process_badges(%{})
{:ok, %{}}

report_badge(badge)

Reports that a badge was posted to Slack by setting a Redis key.

Example

iex> Slacktapped.Badges.report_badge(%{"user_badge_id" => 876})
{:ok, %{"user_badge_id" => 876}}