Slacktapped v1.0.0 Slacktapped.Checkins

Functions

is_eligible_checkin_post(checkin)

Determines if a checkin is eligible to be posted to Slack.

Checkin is ineligible to post if:

There is a Redis key indicating that this checkin was posted under this instance, and had an image with it.

There is a Redis key indicating that this checkin was posted under this instance, and had no image with it, but the checkin we see now still does not have an image.

Examples

A checkin that was not already posted:

iex> Slacktapped.Checkins.is_eligible_checkin_post(%{"checkin_id" => 9985})
{:ok, %{"checkin_id" => 9985}}

A checkin that was already posted, and had an image:

iex> Slacktapped.Checkins.is_eligible_checkin_post(%{"checkin_id" => 9988})
{:error, %{"checkin_id" => 9988}}

A checkin that was already posted and had no image, and we do not currently have an image:

iex> Slacktapped.Checkins.is_eligible_checkin_post(%{"beer" => %{}, "checkin_id" => 5566})
{:error, %{"beer" => %{}, "checkin_id" => 5566}}

A checkin that was already posted and had no image, and we now have an image:

iex> Slacktapped.Checkins.is_eligible_checkin_post(%{
...>   "beer" => %{},
...>   "checkin_id" => 5566,
...>   "media" => %{
...>     "items" => [
...>       %{
...>         "photo" => %{
...>           "photo_id" => 987,
...>           "photo_img_lg" => "http://path/to/beer/image"
...>         }
...>       }
...>     ]
...>   }
...> })
{:ok, %{
  "beer" => %{},
  "checkin_id" => 5566,
  "media" => %{
    "items" => [
      %{
        "photo" => %{
          "photo_id" => 987,
          "photo_img_lg" => "http://path/to/beer/image"
        }
      }
    ]
  }
}}

parse_checkin(checkin)

Parses a checkin into an attachment for Slack.

Examples

A typical checkin with a rating and comment:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "attachments" => [],
...>   "user" => %{
...>     "user_name" => "nicksergeant",
...>     "user_avatar" => "http://path/to/user/avatar"
...>   },
...>   "beer" => %{
...>     "bid" => 123,
...>     "beer_abv" => 4.5,
...>     "beer_label" => "http://path/to/beer/label",
...>     "beer_name" => "IPA",
...>     "beer_slug" => "two-lake-ipa",
...>     "beer_style" => "American IPA"
...>   },
...>   "brewery" => %{
...>     "brewery_id" => 1,
...>     "brewery_label" => "http://path/to/brewery/label",
...>     "brewery_name" => "Two Lake"
...>   },
...>   "checkin_comment" => "Lovely!",
...>   "checkin_id" => 567,
...>   "rating_score" => 3.5
...> })
{:ok,
  %{
    "author_icon" => "http://path/to/user/avatar",
    "author_link" => "https://untappd.com/user/nicksergeant",
    "author_name" => "nicksergeant",
    "color" => "#FFCF0B",
    "fallback" => "nicksergeant is drinking IPA.",
    "footer" => "<https://untappd.com/brewery/1|Two Lake>",
    "footer_icon" => "http://path/to/brewery/label",
    "image_url" => "http://path/to/beer/label",
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/nicksergeant|nicksergeant> is drinking " <>
      "<https://untappd.com/b/two-lake-ipa/123|IPA> " <>
      "(American IPA, 4.5% ABV)." <>
      "\nThey rated it a 3.5 and said:\n>Lovely!\n" <>
      "<https://untappd.com/user/nicksergeant/checkin/567|Toast »>",
    "title" => "IPA",
    "title_link" => "https://untappd.com/b/two-lake-ipa/123"
  }
}

A checkin without a rating:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "attachments" => [],
...>   "checkin_comment" => "Lovely!"
...> })
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV)." <>
      "\nThey said:\n>Lovely!\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin without a comment:

iex> Slacktapped.Checkins.parse_checkin(%{"rating_score" => 1.5})
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV)." <>
      "\nThey rated it a 1.5.\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin without a comment or rating:

iex> Slacktapped.Checkins.parse_checkin(%{})
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV).\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin with a rating of 0:

iex> Slacktapped.Checkins.parse_checkin(%{"rating_score" => 0})
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV).\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin with a rating of 0 and a comment:

iex> Slacktapped.Checkins.parse_checkin(%{
...>  "checkin_comment" => "Yuck.",
...>  "rating_score" => 0
...>})
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV)." <>
      "\nThey said:\n>Yuck.\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin with an image:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "media" => %{
...>     "items" => [
...>       %{
...>         "photo" => %{
...>           "photo_id" => 987,
...>           "photo_img_lg" => "http://path/to/beer/image"
...>         }
...>       }
...>     ]
...>   }
...> })
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => "http://path/to/beer/image",
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV).\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin with a venue:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "attachments" => [],
...>   "venue" => %{
...>     "venue_id" => 789,
...>     "venue_name" => "Venue Name",
...>     "venue_slug" => "venue-slug"
...>   }
...> })
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => nil,
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV) " <>
      "at <https://untappd.com/v/venue-slug/789|Venue Name>.\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

A checkin with an image, with a previously-posted checkin with no image:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "checkin_id" => 5566,
...>   "media" => %{
...>     "items" => [
...>       %{
...>         "photo" => %{
...>           "photo_id" => 987,
...>           "photo_img_lg" => "http://path/to/beer/image"
...>         }
...>       }
...>     ]
...>   }
...> })
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => "http://path/to/beer/image",
    "mrkdwn_in" => ["text"],
    "text" => "",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

The default Untappd beer label is ignored:

iex> Slacktapped.Checkins.parse_checkin(%{
...>   "attachments" => [],
...>   "beer" => %{
...>     "beer_label" => "https://untappd.akamaized.net/site/assets/images/temp/badge-beer-default.png"
...>   }
...> })
{:ok,
  %{
    "author_icon" => nil,
    "author_link" => "https://untappd.com/user/",
    "author_name" => nil,
    "color" => "#FFCF0B",
    "fallback" => " is drinking .",
    "footer" => "<https://untappd.com/brewery/|>",
    "footer_icon" => nil,
    "image_url" => "",
    "mrkdwn_in" => ["text"],
    "text" => "" <>
      "<https://untappd.com/user/|> is drinking " <>
      "<https://untappd.com/b//|> (, % ABV).\n" <>
      "<https://untappd.com/user//checkin/|Toast »>",
    "title" => nil,
    "title_link" => "https://untappd.com/b//"
  }
}

process_checkin(checkin)

Processes a checkin for use in a Slack post:

Verifies the checkin is eligible to post.

Parses the checkin to a Slack attachment.

Adds the attachment to the checkin.

Reports the checkin post to Redis.

report_checkin_type(arg)

Reports that a checkin was posted to Slack by setting a Redis key. There are two possible Redis keys indicating a post was made:

[slacktapped-instance-name]:[checkin_id]:with-image

[slacktapped-instance-name]:[checkin_id]:without-image

Already-posted checkins that were previously posted without an image are posted again indicating that the user added an image to the checkin. If a record exists for the checkin with an image, the checkin is not posted again.

Examples

iex> Slacktapped.Checkins.report_checkin_type({:ok, %{}})
{:ok, %{"reported_as" => "without-image"}}

iex> Slacktapped.Checkins.report_checkin_type({:ok, %{
...>   "media" => %{
...>     "items" => [
...>       %{
...>         "photo" => %{
...>           "photo_id" => 987,
...>           "photo_img_lg" => "http://path/to/beer/image"
...>         }
...>       }
...>     ]
...>   }
...> }})
{:ok, %{
  "reported_as" => "with-image",
  "media" => %{
    "items" => [
      %{
        "photo" => %{
          "photo_id" => 987,
          "photo_img_lg" => "http://path/to/beer/image"
        }
      }
    ]
  }
}}