Slacktapped v1.0.0 Slacktapped.Comments

Functions

is_eligible_comment(comment)

Determines if a comment is eligible to be posted to Slack.

Comment is ineligible if we have already posted it to Slack.

Examples

iex> Slacktapped.Comments.is_eligible_comment(%{"comment_id" => 3344})
true

iex> Slacktapped.Comments.is_eligible_comment(%{"comment_id" => 3355})
false

parse_comment(comment, checkin)

Parses a comment into an attachment for Slack.

Examples

iex> Slacktapped.Comments.parse_comment(
...> %{
...>   "comment_id" => 3593,
...>   "comment" => "Great beer!",
...>   "user" => %{
...>     "first_name" => "George",
...>     "last_name" => "C.",
...>     "uid" => 358,
...>     "user_link" => "http://path/to/comment/user",
...>     "user_name" => "george"
...>   }
...> },
...> %{
...>   "attachments" => [],
...>   "user" => %{
...>     "uid" => 357,
...>     "user_name" => "nicksergeant"
...>   },
...>   "beer" => %{
...>     "bid" => 123,
...>     "beer_name" => "IPA",
...>     "beer_slug" => "two-lake-ipa"
...>   },
...>   "checkin_id" => 567
...> })
%{
  "color" => "#FFCF0B",
  "fallback" => "George C. commented on nicksergeant's checkin.",
  "mrkdwn_in" => ["text"],
  "text" => "<http://path/to/comment/user|George C.> commented on " <>
    "<https://untappd.com/user/nicksergeant/checkin/567|nicksergeant's " <>
    "checkin> of <https://untappd.com/b/two-lake-ipa/123|IPA> and " <>
    "said:\n>Great beer!"
}

If the checkin’s owner posts a comment on their own checkin:

iex> Slacktapped.Comments.parse_comment(
...> %{
...>   "comment_id" => 3593,
...>   "comment" => "I loved this!",
...>   "user" => %{
...>     "uid" => 357,
...>     "user_name" => "nicksergeant"
...>   }
...> },
...> %{
...>   "attachments" => [],
...>   "user" => %{
...>     "uid" => 357,
...>     "user_name" => "nicksergeant"
...>   },
...>   "beer" => %{
...>     "bid" => 123,
...>     "beer_name" => "IPA",
...>     "beer_slug" => "two-lake-ipa"
...>   },
...>   "checkin_id" => 567
...> })
%{
  "color" => "#FFCF0B",
  "fallback" => "nicksergeant commented on nicksergeant's checkin.",
  "mrkdwn_in" => ["text"],
  "text" => "<|nicksergeant> commented on " <>
    "<https://untappd.com/user/nicksergeant/checkin/567|their " <>
    "checkin> of <https://untappd.com/b/two-lake-ipa/123|IPA> and " <>
    "said:\n>I loved this!"
}

process_comment(comment, checkin)

Reports the comment as posted to Redis, and returns a parsed comment as an attachment for Slack.

process_comments(checkin)

Processes a checkins comments.

Loops through all comments and determines if they are eligible to be posted to Slack. If they are, they’re processed and added to the checkins attachments.

Examples

iex> Application.put_env(:slacktapped, :ignore_comments, true)
iex> Slacktapped.Comments.process_comments(%{})
{:error, %{}}

iex> Application.put_env(:slacktapped, :ignore_comments, nil)
iex> Slacktapped.Comments.process_comments(%{})
{:ok, %{}}

report_comment(comment)

Reports that a comment was posted to Slack by setting a Redis key.

Example

iex> Slacktapped.Comments.report_comment(%{"comment_id" => 345})
{:ok, %{"comment_id" => 345}}